libvirt_server/pyproject.toml
[project]
name = "libvirt-server"
version = "0.1.0"
description = "A simple mcp server about libvirt"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argcomplete>=3.6.2",
    "dotenv>=0.9.9",
    "libvirt-python>=11.4.0",
    "lxml>=5.4.0",
    "mcp[cli]>=1.9.4",
    "openai-agents>=0.0.17",
    "paramiko>=3.5.1",
    "prompt-toolkit>=3.0.51",
    "pyyaml>=6.0.2",
    "tqdm>=4.67.1",
]
